API Reference

YAML parser

LibYAML.ParserYAML.parse_yaml — Function

parse_yaml(yaml_str::String; multi::Bool)
parse_yaml(yaml_str::Vector{UInt8}; multi::Bool)

Parse a YAML string or file (or vector of UInt8) into a dictionary, vector or nothing.

If a given YAML document contains a dictionary, the parser returns a dictionary.

If a given YAML document contains just a list of variables, the parser returns a vector.

If a given YAML document contains no information (i.e. empty), the parser returns nothing

or empty dictionary.

Returns a sequence of documents parsed from YAML string given.

Keyword arguments

multi::Bool = false: If YAML is multidocumental, allows the reader to choose between obtaining all documents at once or only the first one.

Examples

julia> yaml_str = """
        name: Alice
        array: 
          - 1
          - 2
          - a: 3 
            b: null
        dict: 
          a: 1
          b: 
            - w
            - d
       """;

julia> parse_yaml(yaml_str)
Dict{String, Any}(
    "dict" => Dict{Any, Any}(
        "b" => Any["w", "d"], 
        "a" => "1"
    ), 
    "name" => "Alice", 
    "array" => Any["1", "2", Dict{Any, Any}("b" => nothing, "a" => "3")]
)

LibYAML.ParserYAML.open_yaml — Function

open_yaml(path::AbstractString; multi::Bool)

Read a YAML file from a given path and parse it.

Keyword arguments

multi::Bool = false: If YAML is multidocumental, allows the reader to choose between obtaining all documents at once or only the first one.

open_yaml(io::IO; multi::Bool)

Reads a YAML file from a given io and parse it.

Keyword arguments

multi::Bool = false: If YAML is multidocumental, allows the reader to choose between obtaining all documents at once or only the first one.